Use uint16_t for view numbers

+++ client/src/lvgl-gui/file_system.h
@@ -9,7 +9,7 @@ typedef struct {
   CardType type;
   char *filename;
   char *dir;
-  uint8_t prev_view;
+  uint16_t prev_view;
   bool leaving;
 } FileManagerData;
